f! added minor changes to allow the minimal example to run
Note: .copy() by default does a shallow copy. Using .deepcopy() is required for a deep copy

the `species` that is passed around was previously a dictionary, not a Species object. This commit fixes that so T3 can now complete 1 iteration. It currently fails on the second iteration due to incorrectly writing the input.py file.